chore: move utils to top level
