+++ src/BootstrapModalManager.js
