+++ src/CarouselCaption.js
@@ -1,4 +1,4 @@
-import createWithBsPrefix from './utils/createWithBsPrefix';
+import createWithBsPrefix from './createWithBsPrefix';
 
 export default createWithBsPrefix('carousel-caption', {
   Component: 'div',
