+++ src/InputGroup.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import React from 'react';
 
-import createWithBsPrefix from './utils/createWithBsPrefix';
+import createWithBsPrefix from './createWithBsPrefix';
 import { useBootstrapPrefix } from './ThemeProvider';
 
 const propTypes = {
