+++ src/ToastBody.js
@@ -1,3 +1,3 @@
-import createWithBsPrefix from './utils/createWithBsPrefix';
+import createWithBsPrefix from './createWithBsPrefix';
 
 export default createWithBsPrefix('toast-body');
