+++ src/createChainedFunction.js
